Decouple RV32 instructions as implementation templates
This commit prepares the groundwork for the future development of
a baseline JIT compiler. By decoupling the RV32 instructions, we
create reusable implementation templates that can be leveraged for
code generation in the JIT compiler.

This decoupling enables straightforward and flexible code generation
strategies, without code duplication between interpreter and MIR
based JIT compiler.
